01_django_server/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
requests==2.31.0

# MySQL
mysqlclient==2.2.0

# DICOM 처리
pydicom==2.4.3

# RabbitMQ 클라이언트
pika==1.3.2

# 기타
python-dotenv==1.0.0
